Creating this second branch after refactoring to the code to run it as a package.
